+++ packages/components/src/components/icon/template.ts
@@ -1,0 +1,10 @@
+import { html } from '@microsoft/fast-element';
+import { Icon } from './class.js';
+
+export const iconTemplate = () => html<Icon>`
+  <span part="root">
+    <span class="icon">
+      <slot></slot>
+    </span>
+  </span>
+`;