+++ packages/components/src/components/index.ts
@@ -1,1 +1,2 @@
+export * from './icon/index.js';
 export * from './logo/index.js';