+++ packages/components/CHANGELOG.md
@@ -21,3 +21,4 @@ The following icons are used in commit messages and this changelog.
 ### ✨ Features
 
 - #18: Basic package setup, linting, dependencies and typescript
+- #18: Added initial `<sds-icon />` component